Adds opacity animation and some SwipeStatus to a separate file.
